# swagger-specs/swagger-ui-dist/personal-rag-server-openapi.yaml
openapi: 3.0.3
info:
    title: Personal RAG Server API
    description: |
        Personal RAG Server with FastAPI - Secure and Production Ready

        Vector Database: Local (ChromaDB-based) with `philosophical_768` default collection
    version: 1.0.0
    contact:
        name: API Support
servers:
    - url: /
      description: Root
    - url: /api/v1
      description: API v1
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
        apiKeyAuth:
            type: apiKey
            in: header
            name: X-API-Key
    schemas:
...
        ReasonerFilters:
            type: object
            properties:
                book_ids:
                    type: array
                    items:
                        type: string
                authors:
                    type: array
                    items:
                        type: string
                chapters:
                    type: array
                    items:
                        type: string
        ReasonerRequest:
            type: object
            properties:
                query:
                    type: string
                user_prompt_template:
                    type: string
                    description: Template with placeholders {context} and {question}
                system_prompt:
                    type: string
                collection_name:
                    type: string
                top_k:
                    type: integer
                    default: 8
                filters:
                    $ref: '#/components/schemas/ReasonerFilters'
            required:
                - query
                - user_prompt_template
        ReasonerResponse:
            type: object
            properties:
                content:
                    type: string
                model:
                    type: string
                retrieved_documents:
                    type: array
                    items:
                        type: object
                        additionalProperties: true
            required:
                - content
...
        Assistant:
            type: object
            properties:
                id:
                    type: string
                    description: Assistant identifier
                object:
                    type: string
                    default: 'assistant'
                    description: Object type
                name:
                    type: string
                    description: Assistant name
                description:
                    type: string
                    nullable: true
                    description: Assistant description
                model:
                    type: string
                    description: LLM model used by assistant
                instructions:
                    type: string
                    nullable: true
                    description: Assistant instructions
                weltanschauung:
                    type: string
                    nullable: true
                    enum: [Idealismus, Materialismus, Realismus, Spiritualismus]
                    description: Philosophical worldview
                status:
                    type: string
                    description: Assistant status
                created_at:
                    type: string
                    description: Creation timestamp
                metadata:
                    type: object
                    additionalProperties: true
                    nullable: true
                    description: Additional metadata
                capabilities:
                    $ref: '#/components/schemas/AssistantCapabilities'
                    nullable: true
                    description: Assistant capabilities
            required:
                - id
                - object
                - name
                - model
                - status
                - created_at
        AssistantCapabilities:
            type: object
            properties:
                templates:
                    type: array
                    items:
                        type: string
                        enum: [resolve, reformulate, glossary]
                max_context_length:
                    type: integer
                supported_languages:
                    type: array
                    items:
                        type: string
                        enum: [de, en]
        AssistantListResponse:
            type: object
            properties:
                object:
                    type: string
                    default: 'list'
                    description: Object type
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/Assistant'
                    description: List of assistants
                first_id:
                    type: string
                    nullable: true
                    description: First assistant ID
                last_id:
                    type: string
                    nullable: true
                    description: Last assistant ID
                has_more:
                    type: boolean
                    default: false
                    description: Whether there are more assistants
            required:
                - object
                - data
                - has_more
...
        AssistantQuery:
            type: object
            properties:
                query:
                    type: string
                template:
                    type: string
                template_variables:
                    type: object
                    additionalProperties: true
            required:
                - query
        AssistantResponse:
            type: object
            properties:
                response:
                    type: object
                    additionalProperties: true
                model:
                    type: string
                weltanschauung:
                    type: string
                retrieved_documents:
                    type: array
                    items:
                        type: object
                        additionalProperties: true
            required:
                - response

        # Phase 4 Template Endpoint Schemas
...
paths:
...
    /rag/reasoner:
        post:
            summary: Retrieve book chunks and call DeepSeek Reasoner with augmented prompt
            operationId: reasoner
            tags:
                - rag
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ReasonerRequest'
            responses:
                '200':
                    description: Reasoner response with retrieved documents
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ReasonerResponse'
                '400':
                    description: Invalid request
                '500':
                    description: Reasoner failed
    /assistants:
        get:
            summary: List all assistants with optional filtering and capabilities
            operationId: list_assistants
            tags:
                - assistants
            security:
                - bearerAuth: []
                - apiKeyAuth: []
            parameters:
                - name: weltanschauung
                  in: query
                  required: false
                  schema:
                      type: string
                      enum:
                          [Idealismus, Materialismus, Realismus, Spiritualismus]
                  description: Filter assistants by weltanschauung
                  example: Idealismus
                - name: status
                  in: query
                  required: false
                  schema:
                      type: string
                      enum: [Ready, Initializing, Failed]
                  description: Filter assistants by status
                  example: Ready
                - name: include_capabilities
                  in: query
                  required: false
                  schema:
                      type: boolean
                      default: false
                  description: Include template capabilities in response
                  example: true
            responses:
                '200':
                    description: List of assistants with capabilities
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AssistantListResponse'
                            example:
                                object: 'list'
                                data:
                                    - id: 'aurelian-i--schelling'
                                      object: 'assistant'
                                      name: 'aurelian-i--schelling'
                                      description: 'Philosophical assistant for Idealism worldview - focuses on ideas as primary reality'
                                      model: 'deepseek-reasoner'
                                      instructions: 'Philosophical assistant with specialized worldview knowledge'
                                      weltanschauung: 'Idealismus'
                                      status: 'Ready'
                                      created_at: '2024-01-15T10:30:00Z'
                                      metadata:
                                          worldview: 'Idealismus'
                                          created_by: 'pinecone-assistant-api'
                                      capabilities:
                                          templates:
                                              [
                                                  'resolve',
                                                  'reformulate',
                                                  'glossary',
                                              ]
                                          max_context_length: 8192
                                          supported_languages: ['de', 'en']
                                first_id: 'aurelian-i--schelling'
                                last_id: 'aurelian-i--schelling'
                                has_more: false
                '401':
                    description: Unauthorized
    /assistants/{assistant_id}:
        get:
            summary: Get assistant details
            operationId: get_assistant
            tags:
                - assistants
            security:
                - bearerAuth: []
                - apiKeyAuth: []
            parameters:
                - name: assistant_id
                  in: path
                  required: true
                  schema:
                      type: string
                  example: 'aurelian-i--schelling'
            responses:
                '200':
                    description: Assistant details
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Assistant'
                '401':
                    description: Unauthorized
                '404':
                    description: Assistant not found

        delete:
            summary: Delete a Pinecone assistant
            operationId: delete_assistant
            tags:
                - assistants
            security:
                - bearerAuth: []
                - apiKeyAuth: []
            parameters:
                - name: assistant_id
                  in: path
                  required: true
                  schema:
                      type: string
                  example: 'aurelian-i--schelling'
            responses:
                '200':
                    description: Assistant deleted successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: string
                                        description: Assistant ID
                                    object:
                                        type: string
                                        default: 'assistant.deleted'
                                        description: Object type
                                    deleted:
                                        type: boolean
                                        default: true
                                        description: Deletion status
                                required:
                                    - id
                                    - object
                                    - deleted
                '401':
                    description: Unauthorized
                '404':
                    description: Assistant not found
                '500':
                    description: Failed to delete assistant
    /assistants/{assistant_id}/query:
        post:
            summary: Query an assistant
            operationId: query_assistant
            tags:
                - assistants
            security:
                - bearerAuth: []
                - apiKeyAuth: []
            parameters:
                - name: assistant_id
                  in: path
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AssistantQuery'
            responses:
                '200':
                    description: Assistant response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AssistantResponse'
                '400':
                    description: Invalid request
                '401':
                    description: Unauthorized
                '404':
                    description: Assistant not found
                '500':
                    description: Query failed

    # Phase 4 Template Endpoints
...